% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitching.R
\docType{data}
\name{cheatsheets_page}
\alias{cheatsheets_page}
\title{importFrom magrittr "\%>\%"}
\format{
An object of class \code{xml_document} (inherits from \code{xml_node}) of length 2.
}
\usage{
cheatsheets_page
}
\description{
importFrom magrittr "\%>\%"
}
\keyword{datasets}
